requirements.txt
aniso8601==8.0.0
asn1crypto==1.3.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
click==7.1.1
cryptography==2.8
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.20.0
flask-marshmallow==0.11.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
fuzzywuzzy==0.17.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.9.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.1
wincertstore==0.2
